wrap importCellParams.js content with ActionDialog
